tree-wide: fix type comparisons

+++ src/proc_fuse.c
@@ -329,7 +329,7 @@ static int proc_swaps_read(char *buf, size_t size, off_t offset,
 	size_t linelen = 0;
 
 	if (offset) {
-		int left;
+		size_t left;
 
 		if (offset > d->size)
 			return -EINVAL;
@@ -426,7 +426,7 @@ static int proc_swaps_read(char *buf, size_t size, off_t offset,
 	d->cached = 1;
 	d->size = (int)total_len;
 
-	if (total_len > size)
+	if ((size_t)total_len > size)
 		total_len = size;
 	memcpy(buf, d->buf, total_len);
 
@@ -500,7 +500,7 @@ static int proc_diskstats_read(char *buf, size_t size, off_t offset,
 	int ret;
 
 	if (offset) {
-		int left;
+		size_t left;
 
 		if (offset > d->size)
 			return -EINVAL;
@@ -631,7 +631,7 @@ static int proc_diskstats_read(char *buf, size_t size, off_t offset,
 		l = snprintf(cache, cache_size, "%s", lbuf);
 		if (l < 0)
 			return log_error(0, "Failed to write cache");
-		if (l >= cache_size)
+		if ((size_t)l >= cache_size)
 			return log_error(0, "Write to cache was truncated");
 
 		cache += l;
@@ -808,7 +808,7 @@ static int proc_uptime_read(char *buf, size_t size, off_t offset,
 #endif
 
 	if (offset) {
-		int left;
+		size_t left;
 
 		if (offset > d->size)
 			return -EINVAL;
@@ -840,7 +840,7 @@ static int proc_uptime_read(char *buf, size_t size, off_t offset,
 
 	d->cached = 1;
 	d->size = total_len;
-	if (total_len > size)
+	if ((size_t)total_len > size)
 		total_len = size;
 	memcpy(buf, d->buf, total_len);
 
@@ -873,7 +873,7 @@ static int proc_stat_read(char *buf, size_t size, off_t offset,
 	int cg_cpu_usage_size = 0;
 
 	if (offset) {
-		int left;
+		size_t left;
 
 		if (offset > d->size)
 			return -EINVAL;
@@ -947,7 +947,7 @@ static int proc_stat_read(char *buf, size_t size, off_t offset,
 			l = snprintf(cache, cache_size, "%s", line);
 			if (l < 0)
 				return log_error(0, "Failed to write cache");
-			if (l >= cache_size)
+			if ((size_t)l >= cache_size)
 				return log_error(0, "Write to cache was truncated");
 
 			cache += l;
@@ -984,7 +984,7 @@ static int proc_stat_read(char *buf, size_t size, off_t offset,
 			l = snprintf(cache, cache_size, "cpu%d%s", curcpu, c);
 			if (l < 0)
 				return log_error(0, "Failed to write cache");
-			if (l >= cache_size)
+			if ((size_t)l >= cache_size)
 				return log_error(0, "Write to cache was truncated");
 
 			cache += l;
@@ -1016,7 +1016,7 @@ static int proc_stat_read(char *buf, size_t size, off_t offset,
 				     cg_cpu_usage[physcpu].system, new_idle);
 			if (l < 0)
 				return log_error(0, "Failed to write cache");
-			if (l >= cache_size)
+			if ((size_t)l >= cache_size)
 				return log_error(0, "Write to cache was truncated");
 
 			cache += l;
@@ -1160,7 +1160,7 @@ static int proc_meminfo_read(char *buf, size_t size, off_t offset,
 	int ret;
 
 	if (offset) {
-		int left;
+		size_t left;
 
 		if (offset > d->size)
 			return -EINVAL;
@@ -1370,7 +1370,7 @@ static int proc_meminfo_read(char *buf, size_t size, off_t offset,
 		l = snprintf(cache, cache_size, "%s", printme);
 		if (l < 0)
 			return log_error(0, "Failed to write cache");
-		if (l >= cache_size)
+		if ((size_t)l >= cache_size)
 			return log_error(0, "Write to cache was truncated");
 
 		cache += l;
@@ -1402,7 +1402,7 @@ static int proc_slabinfo_read(char *buf, size_t size, off_t offset,
 	pid_t initpid;
 
 	if (offset) {
-		int left;
+		size_t left;
 
 		if (offset > d->size)
 			return -EINVAL;
@@ -1439,7 +1439,7 @@ static int proc_slabinfo_read(char *buf, size_t size, off_t offset,
 		ssize_t l = snprintf(cache, cache_size, "%s", line);
 		if (l < 0)
 			return log_error(0, "Failed to write cache");
-		if (l >= cache_size)
+		if ((size_t)l >= cache_size)
 			return log_error(0, "Write to cache was truncated");
 
 		cache += l;
